Add an exemplary github actions workflow
The added workflow (`.github/workflows/ci_rp.yml`) will build the SDK
base docker image, and on top of that the SDK RP docker image.  Once
done, that docker image is run as a GH CI action.  The image contains
a shell script (`docker/rp/run_rp_test.sh`) which run an RP smoke test
in that container (the ultimate goal is to run unit tests there).

The action has been registered in GitHub, and will run whenever a commit
to master happens, or whenever a PR sees a new commit.  The badge can be
used in MD files (see `README.md`) to document the result of the action.

Co-authored-by: Stephen Herbein <SteVwonder@users.noreply.github.com>