dockerfiles: unify virtualenv for python deps
Problem: radical pilot sets up a virtualenv to install dependencies in,
parsl installs into the system directory, and flux depends on yum-installed
dependencies.

Unify python dependencies into a single virtualenv.  Parsl and RP both
install via pip, so create and activate the virtualenv in the base image.
Flux uses m4 to detect the python version, and it start by looking for
a `python3.6` binary, which is not in the virtualenv and only in
`/usr/bin`.  So explicitly set `PYTHON_VERSION=3` when configuring flux
to properly detect the virtualenv python.