base: manually compile OpenMPI
Problem: the OpenMPI packaged with centos 7 does not include the PMI
plugin, which is required to boot under Flux. The OpenMPI version
packaged with centos 8 (4.0.5) has a bug that causes a segfault on
finalization, which is fixed in the next patch version.
PR that fixes the bug: (https://github.com/open-mpi/ompi/pull/8380)

Solution: for centos 7, hand compile the same version of OpenMPI (1.10),
but with the `--with-pmi` flag.  For centos 8, hand compile the version
of OpenMPI with the patch (4.0.6).  Since this is an involved process,
add a helper script that will also make adding other MPIs much easier.