flux: add job-shell plugin to fix pmi.clique
Problem: When running with `flux start --test-size`, as the flux integration
tests are, openmpi hangs. This is because when Flux is started with
`--test-size`, the job-shell setting `pmi.clique` defaults to singleton, which
means Flux's PMI tells OpenMPI that all ranks are on the same node, even when
you use `flux mini run -N2`. This causes OpenMPI to hang.

Solution: Add a job-shell plugin that forces Flux's PMI to "lie" to OpenMPI
that each shell is running on a different node. OpenMPI properly bootstraps
under this config.